Move logging-related items to mf_logging.
