+++ metricflow/engine/metricflow_engine.py
@@ -31,7 +31,7 @@ from metricflow.execution.execution_plan import ExecutionPlan, SqlQuery
 from metricflow.execution.execution_plan_to_text import execution_plan_to_text
 from metricflow.execution.executor import SequentialPlanExecutor
 from metricflow.filters.time_constraint import TimeRangeConstraint
-from metricflow.formatting import indent_log_line
+from metricflow.mf_logging.formatting import indent_log_line
 from metricflow.model.semantic_manifest_lookup import SemanticManifestLookup
 from metricflow.model.semantics.linkable_element_properties import (
     LinkableElementProperties,
