+++ metricflow/mf_logging/__init__.py
@@ -1,0 +1,4 @@
+"""Contains logging-related classes / functions.
+
+Named with the 'mf_' prefix to avoid collision with the built-in module.
+"""
