+++ metricflow/query/issues/group_by_item_resolver/invalid_use_of_date_part.py
@@ -5,7 +5,7 @@ from typing import Sequence, Tuple
 
 from typing_extensions import override
 
-from metricflow.collection_helpers.pretty_print import mf_pformat
+from metricflow.mf_logging.pretty_print import mf_pformat
 from metricflow.query.group_by_item.resolution_path import MetricFlowQueryResolutionPath
 from metricflow.query.issues.issues_base import (
     MetricFlowQueryIssueType,
