+++ metricflow/test/collection_helpers/test_pretty_print.py
@@ -6,8 +6,8 @@ import textwrap
 from dbt_semantic_interfaces.implementations.elements.dimension import PydanticDimension
 from dbt_semantic_interfaces.type_enums import DimensionType
 
-from metricflow.collection_helpers.pretty_print import mf_pformat, mf_pformat_many
-from metricflow.formatting import indent_log_line
+from metricflow.mf_logging.formatting import indent_log_line
+from metricflow.mf_logging.pretty_print import mf_pformat, mf_pformat_many
 from metricflow.test.time.metric_time_dimension import MTD_SPEC_DAY
 
 logger = logging.getLogger(__name__)
