+++ metricflow/test/fixtures/sql_clients/base_sql_client_implementation.py
@@ -9,7 +9,7 @@ import pandas as pd
 from dbt_semantic_interfaces.pretty_print import pformat_big_objects
 
 from metricflow.dataflow.sql_table import SqlTable
-from metricflow.formatting import indent_log_line
+from metricflow.mf_logging.formatting import indent_log_line
 from metricflow.protocols.sql_client import (
     SqlClient,
 )
