+++ metricflow/test/snapshot_utils.py
@@ -11,12 +11,12 @@ import tabulate
 from _pytest.fixtures import FixtureRequest
 from dbt_semantic_interfaces.pretty_print import pformat_big_objects
 
-from metricflow.collection_helpers.pretty_print import mf_pformat
 from metricflow.dag.mf_dag import MetricFlowDag
 from metricflow.dataflow.dataflow_plan import DataflowPlan
 from metricflow.dataflow.dataflow_plan_to_text import dataflow_plan_as_text
 from metricflow.execution.execution_plan import ExecutionPlan
 from metricflow.execution.execution_plan_to_text import execution_plan_to_text
+from metricflow.mf_logging.pretty_print import mf_pformat
 from metricflow.model.semantics.linkable_spec_resolver import LinkableElementSet
 from metricflow.naming.object_builder_scheme import ObjectBuilderNamingScheme
 from metricflow.protocols.sql_client import SqlClient
